Simplified ASCII upper/lower functions and tried to make better use of iterators.
`from_utf8()` used the built-in check to test if a character was valid ascii, but that only checks if the character is within the 7-bit range. We need to use the full 8-bit ascii range for our ascii characters.

`lower()` & `upper()` did not do an in-place update of each character and instead allocated a new vector. I've updated it so it now does the operation in-place.

`is_numeric()`, `is_alphanumeric()`, `is_lower_alphanumeric()`, & `is_upper_alphanumeric()` all had large match statements. I've simplified them with range notation and made use of iterators. This should make it a little more readable and self-explanatory.