Created basic `Base64`, `Base32`, `ExtendedBase32`, and `Base16` types
Each of these types will be used for encoding/decoding various byte streams for DNS. They all use padding and have pre-defined alphabets. The internal representation of each of these types if the raw bytes so there is only encoding/decoding cost if the streams need to be read/written as ascii.

The `BaseConversions` trait is used to convert between each of these types and to/from bytes.