Designed the network serialization/deserialization system
`ToWire` is used to serialize types.

`FromWire` is used to deserialize types.

`CompressionMap` is used during serialization to keep track of domain name pointers and will be used by CDomainName to decompress the data.

`ReadWire` and `WriteWire` types are used to read in bytes our write out bytes. These structs provide a helpful abstraction to ensure that bytes are written correctly and don't overflow the buffer.

Numerous primitive and common types have had their implementations written so that they can be referenced easily. For example, instead of writing a custom handler when a u64 needs to be deserialized, the implementation already exists and can be used.