Properly fill in blank RClass and TTL fields and abstracted `ResourceRecord` more.
Just like how blank domain names are treated, blank TTL or RClass fields are filled in with whatever the value of the last defined one was.

Created a new version of `ResourceRecord` for the `Tokenizer` that abstracts the `ResourceRecord` used by the `Entry`. This version does not allow for any of the required entries to be `Optional`. This will make parsing the resulting records much easier since they are always guaranteed to not be None.