Cleaned up the zone file reader, renamed several parts of the tokenizer, fixed a few bugs, & added functionality to load included files.
File name changes:
1. `text_tokens.rs` => `raw_literals.rs`
2. `entry_text_tokens.rs` => `raw_entries.rs`
3. `entry.rs` => `token_entries.rs`
4. `read_resource_records.rs` => `zone_file_reader.rs`
Most of these files also had their underlying structs/enums names changed as well to match the new names. These names hopefully better match the progression of the parsing stages.

Origin tokens are now treated as a special case, which can appear in various locations. This is needed because we remove the differences between quoted strings and non-quoted strings. If we tried to check if a token was `@` after that point, it would register true even if the token had been quoted (which is not the origin). This helps to solve bugs related to parsing/replacing the origin tokens.

Several of the if-else chains in `tokenizer.rs` that were used to fill in empty values or store the last value for entries were simplified to match statements. This should help make the logic more straightforward.

Logic was added to the main cache which allows it to process `$INCLUDE` entries from a zone file. Now, it will recursively load the included files as sub-iterators, consuming them before continuing with the parent.
WARNING: it does not catch circular inclusions which could result in infinite recursion.