Updated the Well-Known-Ports "e3consultants" was changed to "lsa-comm" for port 3157.
Side Note: I think it may be better to switch this to a configuration file when I get a chance. Although I like the idea of the compiler optimizing the match statement, it makes it harder to configure well-known-ports on the user side. It also means that I have to update the application whenever IANA makes a changes to the port registry. It might be worth creating a config file that the application can parse on startup / during runtime.