Started to design `Bytes` struct to replace or augment the `ReadWire`.
Working with `ReadWire` does not tend to result in the most readable code. I am hoping that I will be able to use this struct to replace `ReadWire`. Since it has a cleaner implementation and all the functions have clear names, it should result in much more readable code.

The other benefit it should provide is that it provides a zero-copy representation of bytes. So, as an example, when bytes are read in from a socket, they can be quickly parsed and broken apart without the overhead that comes with repeatedly copying the stream of bytes. We might even be able to take this one step further and use `Bytes` in structs like `AsciiString`, `Base16`, `Base32`, etc. so that the underlying streams of bytes are almost never copied unless explicitly required.

I am a little concerned that this may inadvertently increase the memory footprint by storing large vectors of bytes, of which only a small portion is actually being referenced.