Switched several types to use `TinyVec` and `ArrayVec` to reduce the number of heap allocations required (especially during serialization and deserialization).
AsciiStrings are normally used in CDomainNames and DomainNames. In some simple benchmarking, I found that setting the array size to 5 could result in an around 25% performance boost for deserialization. In the future, I may make this number configurable through a generic type to allow other types to optimize for their specific use cases.