Fixed bugs that could allow UDP queries to wait forever and all socket readers to loop infinitely.
One of the `select!` statements for the UDP query manager did not have a timeout or cancellation token being awaited. This could (and did) result in some queries which failed to receive responses via UDP never exiting or not exiting for a long time if a TCP connection could not be established. To fix this, it now waits on both a cancellation token and uses the TCP timeout. This bounds how long it is allowed to wait before killing the query and ensures that the query will shutdown when a cancellation token is used.

A poor decision in the error handling for both the TCP and UDP listeners allowed the listener to loop infinitely in the default error case. So, if an error was not getting handled, the loop would never exit. When hundreds of connections are being handled, this can quickly spike CPU usage to 100% if a few of those connections have an unhandled error. To solve, all errors have been made fatal.

Also, UnexpectedEof was being handled as a non-fatal error. This was a misunderstanding of the documentation on my part. That should have always been fatal.