Bug Fixes to `MixedSocket` and made TCP
Establishing cancelable.
The listener tasks were being spawned before the state was transitioned to `Managed`. This violated one the requirements of the state machine and could have resulted in issues.

Added a cancellation token to the initialization stage of the TCP socket. this way, if a connection is being initialized and a shutdown of the socket is requested, the connection request can be cancelled without breaking the state machine.

There still seems to be issues with client not shutting down cleanly. I am not sure why yet but I suspect there is some bug in this network library I haven't stumbled across yet.