Returned to a shared query system A lot of issues seem to stem from each query being individually owned and polled by the one who created it. This is because they can cancel that task at any time, and any other task that was waiting for the response for that query will be left waiting (if it didn't get far enough along). So, the best way to solve that appears to make the queries themselves separate from the task which created the query. The queries are now polled by the runtime. This way, any task can ask a question and the query is guaranteed to be polled until it completes or errors and the result will always be sent to any listeners. In tests, this change has resulted in significantly more successful queries that I've achieved in the past when making 100,000+ concurrent queries. Additionally, it does not seem to halt which is a good indication that I haven't created any deadlocks.
In a future commit, I may need to allow the socket to write results directly into the cache so that queries that succeed after they are abandoned can still contribute to the global cache.

Error handling can probably be greatly improved by wrapping the messages being sent in some sort of error type. This is probably a change I will make in the future.

A decent number of comments were also not updated in this commit. Those are also going to need to be cleaned up when I have the time and energy.