Further improved round robin query by not using sorting to determine the fastest name server or address Instead of sorting the fastest name servers and popping the fastest from the end of a vector, the algorithm can make use of iterators and take the fastest name server directly from a vector. The sames goes for selecting the fastest address.
While making these changes, also fixed a minor logic error when selecting the address for a network query. Here, the first address selected is not based on the sorted addresses, so it may not be fastest.

The processes for polling multiple concurrent queries does not require a loop because timeouts only need to be polled once, even if the timeout is reset. This reduces the complexity of the poll function.