Modified the domain name representation to allow for reverse iteration Added an additional vector with length octets to each domain name. This allows for reverse iteration over the domain name by providing a way to determine how many octets are in each label starting from the back. This allows the internal representation to remain the same as the wire format which will always have the fastest serialization / deserialization time. Since most common domain names don't have 14 labels, the use of a TinyVec to store the additional length octets performed slightly better then using a regular Vec while also not using quite as much memory.
The OwnedLabel also uses a TinyVec, although that has not really been benchmarked for performance. I suspect this is a good place to use it, given that it is not quite as common for an individual label in a domain to be more than 14 characters, but that really should be tested at some point.

I alternatively tried to implement this by adding additional length octets into the primary vector, but this always resulted in a higher performance penalty because serialization and deserialization are significantly faster when the internal representation is the same as the wire format.