Removed the async locks and mutexes from the sockets where it made sense to
Throughout this application, I have stuck to using async locks and mutexes, even when it is not recommended. I am now starting to go back and correct some of those, especially in places where a lock should not be held for more than a few nano-seconds. This will also help to simplify some of the state machines and probably speed up the application since tasks will not have so many yield points.