Improved the state machine of  ActiveQuery and removed the async RwLock
Throughout this application, I have stuck to using async locks and mutexes, even when it is not recommended. I am now starting to go back and correct some of those, especially in places where a lock should not be held for more than a few nano-seconds. This will also help to simplify some of the state machines.

In this particular case, there were also some issues with the state machine that were causing queries to panic. Te improvements to the state machine also correct those panics since RoundRobin is no longer able to be polled after it completes.