Turned the Label Ref into a slice-like type The original intent of the case-insensitive and case-sensitive labels was to create a slice-like type that could reference data in a domain name or an owned label. This would allow domain names and owned labels to be `borrow`ed and would work great with HashMaps. But I didn't quite know how to do that yet so I made the type as a regular struct that contained a reference back into the domain name and owned label. I have since learned a little more and understand how to make this happen better so I have converted them into their slice-like types. I have not yet gone through and optimized the code to make good use of these new types yet.
I have not yet verified that what I am doing is truly safe. I believe it is and most of what I have read in the standard library seems to indicate that this is the correct way to do this, but I left a TODO to make sure I eventually go back and double check that.