Implemented compile-time assertions
The code for this was taken almost entirely from code I wrote for another project, `rsunk`. This implementation uses only standard features and attempts to stay as close to the `std` and `core` implementations as possible, The documentation is also written to mimic that of the `std` and `core` libraries.